+++ oak_functions/experimental/metrics/Cargo.toml
@@ -7,9 +7,14 @@ license = "Apache-2.0"
 
 [dependencies]
 anyhow = "*"
+bincode = "*"
 log = "*"
 maplit = "*"
+oak_functions_abi = { path = "../../abi" }
+oak_functions_extension = { path = "../../extension" }
 oak_functions_util = { path = "../../util" }
 oak_logger = { path = "../../logger" }
 rand = "*"
 serde = { version = "*", features = ["derive"] }
+# Use wasmi in `no_std` mode.
+wasmi = { version = "*", default-features = false, features = ["core"] }
